refactor(obj): move logic target
