+++ lib/CMakeLists.txt
@@ -5,3 +5,4 @@ add_subdirectory(about)
 add_subdirectory(macro)
 add_subdirectory(builtins)
 add_subdirectory(components)
+add_subdirectory(link)
