+++ lib/link/bytes.cpp
