+++ lib/link/bytes.hpp
