+++ lib/link/memmap.cpp
